# Conflicts:
#	CMakeLists.txt
#	compile.sh
#	includes/rt_opencl_params_defines.h
#	includes/rt_parsing.h
#	scenes/texture_scene.json
#	sources/rt_parsing/hardcoded_scene.c
#	sources/rt_parsing/json_parser/rt_json_utils.c
#	sources/rt_parsing/json_parser/rt_parse_string.c

+++ sources/kernel/pathtrace_algo/texture_shade_pathtrace.cl
@@ -4,7 +4,7 @@
 
 float3		texture_shade_pathtrace(
 		__global const t_texture_info *texture_info,
-		__global const float *texture_list,
+		__global const int *texture_list,
 		__global const t_object *object,
 		t_ray *ray,
 		t_rayhit *hit,
